refactor(fspy-shm): split the sides apart over a shared layout
The single protocol module mixed three audiences: writer code, reader
code, and the definitions both must agree on. Split along that line:

- writer.rs — claim, fill, finish
- reader.rs — close, completeness, iteration, and the borrow argument
- layout.rs — only what both sides share: the shape math, the header,
  the slot codec, the ordering contract, and the views struct, renamed
  SharedState -> MappedLayout: it holds no state — it is the layout
  bound to one concrete mapping. Endpoints store it as 'mapped'.
- mod.rs — the surface, overview docs, and integration tests
