+++ crates/fspy_shared/src/ipc/channel/shm_io/mod.rs
@@ -30,8 +30,9 @@
 //! from the returned old values. A failed claim sets the CLOSED gate as
 //! its loss report and leaves the counters bumped, harmlessly: readers
 //! clamp to the region capacities, and committed descriptors are
-//! self-describing ([`layout`]), so the counters never locate data. Every slot has a fixed location, so
-//! an unfinished frame can never hide a later one.
+//! self-describing ([`layout`]), so the counters never locate data. Every
+//! slot has a fixed location, so an unfinished frame can never hide a
+//! later one.
 //!
 //! # Frame lifecycle
 //!
@@ -72,18 +73,23 @@
 //! checks, heartbeats, or timeouts.
 
 mod layout;
+mod reader;
+mod writer;
 
-use std::{
-    fmt,
-    num::NonZeroUsize,
-    ops::{Deref, DerefMut},
-    ptr::{NonNull, slice_from_raw_parts_mut},
-    slice,
-    sync::atomic::{AtomicU64, Ordering},
-};
+use std::ptr::slice_from_raw_parts_mut;
+#[cfg(target_os = "linux")]
+use std::sync::atomic::Ordering;
 
 use fspy_shm::Mapping;
-use layout::{CLOSED, Header};
+#[cfg(target_os = "linux")]
+use layout::MappedLayout;
+// Only tests name the error types; production matches on `Ok`/`Err` alone.
+#[cfg(test)]
+pub use reader::ProtocolError;
+pub use reader::ShmReader;
+#[cfg(test)]
+pub use writer::ClaimError;
+pub use writer::ShmWriter;
 
 // The region arithmetic in `layout` relies on `usize` accommodating sums of
 // 32-bit-bounded quantities, and the descriptor protocol on native 64-bit
@@ -114,598 +120,6 @@ pub fn is_supported_region_len(len: usize) -> bool {
     (layout::HEADER_LEN..=layout::MAX_MAPPING_LEN).contains(&len)
 }
 
-// --- The region views and the ordering contract ----------------------------
-// `SharedState::new` builds three typed views of the region — the
-// `repr(C)` `Header`, the descriptor table as a slice of atomics sized by
-// the mapping length, and the untyped payload area as a raw slice — once,
-// when an endpoint attaches; the endpoint stores them beside the mapping
-// they point into. Every access after that is a plain field access or a
-// bounds-checked index. The payload area stays raw because writers hold
-// exclusive `&mut` borrows into it, which must not alias any shared
-// reference.
-//
-// # Shared atomics
-//
-// The header holds two independent monotonic `AtomicU64` counters:
-//
-// - the **claim counter**: bit 63 is the CLOSED gate, the low bits count
-//   claims ever attempted. Claiming is one wait-free `fetch_add`; the
-//   returned old value carries the claim's slot index, the gate, and — by
-//   comparison against the fixed table capacity — the capacity verdict.
-//   The gate is set by the receiver at close and by every failed claim:
-//   one bit is both the loss report completeness derives from and the
-//   valve that stops writers spending work on a channel whose result the
-//   receiver must already reject.
-// - the **payload counter**: payload bytes ever reserved, bumped by another
-//   wait-free `fetch_add`.
-//
-// Failed claims leave the counters bumped; that is harmless, because the
-// receiver clamps instead of trusting the counts, and committed
-// descriptors carry their own offset and length, so the counters never
-// locate data. The payload counter can even wrap on a long-condemned
-// channel — still harmless: wrapping requires prior failures, failures
-// set the gate, and the gate refuses every claim before a span is built.
-//
-// # Memory-ordering contract
-//
-// Three synchronization rules cover the whole protocol:
-//
-// 1. **Claim versus close** — the receiver's close boundary is a plain
-//    snapshot load of the claim counter: claims ordered at or before the
-//    value it reads (in the counter's modification order) are in the
-//    snapshot; later ones receive slot indices the receiver never visits.
-//    Claims publish no payload data, so `Relaxed` suffices throughout.
-//    The CLOSED gate is not itself the boundary — it stops stragglers
-//    from claiming (and allocating pages) forever; any claim admitted
-//    between the snapshot and the gate lands beyond the snapshot and is
-//    never observed. Completeness rides the same modification order: a
-//    failed claim sets the gate as its loss report before the writer
-//    performs the operation whose record was lost — so the snapshot
-//    either sees the bit, or the loss belongs to an operation performed
-//    after the boundary. A writer that skipped because it saw the bit is
-//    covered the same way: the bit that made it skip either reaches the
-//    snapshot or postdates the boundary. A writer that dies before
-//    setting the bit never performed its operation, so nothing was
-//    actually lost.
-// 2. **Writer commit** — the slot compare-and-swap in `FrameMut::finish`
-//    uses `Release`: every payload write happens-before the committed
-//    descriptor becomes visible.
-// 3. **Receiver observation** — the freeze compare-and-swap in
-//    `ShmReader::close` uses `Acquire` on failure: observing a committed
-//    descriptor also makes the payload writes it published visible, so the
-//    borrows `ShmReader` later hands out read settled bytes.
-
-/// The typed views of the region: the header, the descriptor table
-/// sized from the mapping length, and the raw payload area. Built once
-/// when an endpoint attaches and stored in it; the views stay valid
-/// because they point into the mapping's stable target, not into the
-/// endpoint value.
-#[derive(Clone, Copy)]
-struct SharedState {
-    header: NonNull<Header>,
-    table: NonNull<[AtomicU64]>,
-    payloads: *mut [u8],
-    /// The real mapping length. Not derivable from the parts above: the
-    /// payload region rounds down to whole `u64`s, and re-deriving the
-    /// layout from a shortened length could shift the table boundary.
-    len: usize,
-}
-
-impl SharedState {
-    /// Builds the typed views of a shared mapping.
-    ///
-    /// # Safety
-    ///
-    /// - `mem` must be valid for reads and writes, and its address stable,
-    ///   for as long as the returned views (and any copy of them) are
-    ///   used.
-    /// - The memory must have been zero-initialized when the region was
-    ///   created, and accessed only through this protocol since.
-    ///
-    /// # Panics
-    ///
-    /// Panics when the mapping cannot host the protocol at all: base not
-    /// `u64`-aligned, smaller than the header, or larger than
-    /// [`layout::MAX_MAPPING_LEN`]. These indicate a broken caller, not
-    /// runtime data; senders guard untrusted mappings with
-    /// [`is_supported_region_len`] first.
-    #[expect(clippy::cast_ptr_alignment, reason = "the base is asserted `u64`-aligned below")]
-    unsafe fn new(mem: *mut [u8]) -> Self {
-        let base = mem.cast::<u8>();
-        let len = mem.len();
-        assert!(!base.is_null());
-        assert!(base.addr().is_multiple_of(align_of::<Header>()));
-        assert!((layout::HEADER_LEN..=layout::MAX_MAPPING_LEN).contains(&len));
-        // SAFETY: the base is non-null (asserted), and the header and the
-        // table lie inside the mapping — the header by the asserts above,
-        // the table by `layout::max_slots` — at `u64`-aligned offsets. The
-        // payload area keeps the rest of the mapping as a raw slice;
-        // `layout` bounds every span carved from it.
-        unsafe {
-            Self {
-                header: NonNull::new_unchecked(base.cast::<Header>()),
-                table: NonNull::slice_from_raw_parts(
-                    NonNull::new_unchecked(base.add(layout::HEADER_LEN).cast::<AtomicU64>()),
-                    layout::max_slots(len),
-                ),
-                payloads: std::ptr::slice_from_raw_parts_mut(
-                    base.add(layout::payload_base(len)),
-                    layout::payload_region_len(len),
-                ),
-                len,
-            }
-        }
-    }
-
-    /// The protocol header.
-    const fn header(&self) -> &Header {
-        // SAFETY: `new`'s contract keeps the target valid while any view
-        // is used, and the header consists of atomics, so the shared
-        // borrow is valid even while other threads and processes access
-        // the same memory through them.
-        unsafe { self.header.as_ref() }
-    }
-
-    /// The descriptor table.
-    const fn table(&self) -> &[AtomicU64] {
-        // SAFETY: as for `header`.
-        unsafe { self.table.as_ref() }
-    }
-
-    /// Base address of the region: the header sits at offset zero.
-    const fn base(&self) -> *const u8 {
-        self.header.as_ptr().cast()
-    }
-}
-
-// --- The writer side: claim, fill, finish ----------------------------------
-
-/// A concurrent shared-memory frame writer.
-///
-/// Safe to use across threads and processes at the same time: frames are
-/// reserved with atomic operations, filled in uniquely owned payload spans,
-/// and published with an atomic commit (see the ordering contract above).
-pub struct ShmWriter<M> {
-    /// Owns the region the views point into; dropped with the writer.
-    #[cfg_attr(any(not(test), miri), expect(dead_code, reason = "held to keep the region alive"))]
-    mem: M,
-    state: SharedState,
-}
-
-// SAFETY: the writer touches the region only through the protocol's
-// atomics, which synchronize access from any thread; the stored views
-// point into the mapping's stable, independently owned target, not into
-// the writer value itself.
-unsafe impl<M: Send> Send for ShmWriter<M> {}
-// SAFETY: see the `Send` impl; the writer's shared-reference API is
-// internally synchronized by the protocol.
-unsafe impl<M: Sync> Sync for ShmWriter<M> {}
-
-/// Why a frame could not be claimed.
-#[derive(thiserror::Error, Clone, Copy, PartialEq, Eq, Debug)]
-pub enum ClaimError {
-    /// The CLOSED gate was set: the receiver closed the channel, or an
-    /// earlier failed claim condemned it. Skipping the record is sound
-    /// either way — it is outside the receiver's boundary, or the same
-    /// bit already makes the receiver report the channel incomplete.
-    #[error("the channel has been closed")]
-    Closed,
-    /// The claim was refused for space: the region was full, or the frame
-    /// was larger than the `i32::MAX`-byte frame limit. The loss is
-    /// already recorded — this claim set the CLOSED gate — so the channel
-    /// will report itself incomplete and refuse further claims.
-    #[error("no space left in the shared-memory region")]
-    Capacity,
-}
-
-impl<M: AsRawSlice> ShmWriter<M> {
-    /// Creates a writer backed by a shared-memory region.
-    ///
-    /// # Safety
-    ///
-    /// - `mem.as_raw_slice()` must return a stable, valid pointer to the
-    ///   whole region for the writer's lifetime.
-    /// - The region must have been zero-initialized when it was created and
-    ///   accessed only through this protocol since.
-    ///
-    /// # Panics
-    ///
-    /// Panics when the region is not `u64`-aligned or its size is outside the
-    /// supported range (see [`SharedState::new`]).
-    pub unsafe fn new(mem: M) -> Self {
-        // SAFETY: forwarded from this function's contract, which keeps the
-        // region valid and protocol-governed for the writer's lifetime —
-        // and so for every use of the views, which are stored in and
-        // dropped with the writer.
-        let state = unsafe { SharedState::new(mem.as_raw_slice()) };
-        Self { mem, state }
-    }
-
-    /// Whether the CLOSED gate is set: the receiver closed the channel,
-    /// or an earlier failed claim condemned it.
-    pub fn is_closed(&self) -> bool {
-        self.state.header().claims.load(Ordering::Relaxed) & CLOSED != 0
-    }
-
-    /// Claims a frame of exactly `frame_size` bytes.
-    ///
-    /// The frame is invisible to the receiver until [`FrameMut::finish`]
-    /// commits it. Dropping the frame without finishing abandons the claim:
-    /// the receiver ignores the slot, exactly as if the writer had died.
-    /// Frames larger than `i32::MAX` bytes are refused as
-    /// [`ClaimError::Capacity`].
-    ///
-    /// Wait-free: two `fetch_add`s, no retry loop (rule 1). A claim that
-    /// does not fit fails after setting the CLOSED gate: the receiver
-    /// learns a record was lost, and later claims are refused — their
-    /// records would ride a result the receiver must already reject.
-    pub fn claim_frame(&self, frame_size: NonZeroUsize) -> Result<FrameMut<'_>, ClaimError> {
-        let state = self.state;
-        let payload_len = frame_size.get();
-
-        // Reports that this claim's record was lost, before the writer
-        // moves on (rule 1): the gate makes the receiver report the
-        // channel incomplete, and condemns further claims.
-        let report_loss = || {
-            state.header().claims.fetch_or(CLOSED, Ordering::Relaxed);
-            ClaimError::Capacity
-        };
-
-        // No descriptor can describe a payload this long; refuse it before
-        // touching the counters, so the channel keeps working for every
-        // record after it.
-        if payload_len > layout::MAX_PAYLOAD_LEN {
-            return Err(report_loss());
-        }
-        let reserved_len = layout::reserved_payload_len(payload_len);
-
-        // Payload bytes first, so a payload-capacity failure does not burn a
-        // slot. A failed reservation stays counted — overshoot is harmless
-        // because the counter is not what locates payloads (descriptors are)
-        // and a `u64` cannot realistically wrap.
-        let payload_start =
-            state.header().payload_reserved.fetch_add(reserved_len as u64, Ordering::Relaxed);
-        // Checked: a foreign scribble of the counter must fail the claim,
-        // not wrap the bound into an out-of-bounds reservation.
-        let payload_end = payload_start.checked_add(reserved_len as u64);
-        if payload_end.is_none_or(|end| end > state.payloads.len() as u64) {
-            return Err(report_loss());
-        }
-        let payload_start = usize::try_from(payload_start).expect("bounded by the payload region");
-
-        let claims = state.header().claims.fetch_add(1, Ordering::Relaxed);
-        if claims & CLOSED != 0 {
-            // Not a loss: a record refused after close describes an
-            // operation performed outside the channel's boundary.
-            return Err(ClaimError::Closed);
-        }
-        let slot_index = usize::try_from(claims).expect("claim count exceeds usize");
-        if slot_index >= state.table().len() {
-            return Err(report_loss());
-        }
-
-        // SAFETY: the claim reserved
-        // `[payload_start, payload_start + reserved_len)` — inside the
-        // payload region by the capacity check above — exclusively for this
-        // frame: other writers reserve disjoint spans, and the receiver
-        // never reads a payload before observing its committed descriptor,
-        // which `finish` publishes only when it consumes this borrow.
-        let content = unsafe {
-            slice::from_raw_parts_mut(state.payloads.cast::<u8>().add(payload_start), payload_len)
-        };
-        Ok(FrameMut {
-            state,
-            slot_index,
-            descriptor: layout::committed(
-                layout::payload_base(state.len) + payload_start,
-                payload_len,
-            ),
-            content,
-        })
-    }
-
-    // Unwrap `self` and return the underlying memory.
-    #[cfg(all(test, not(miri)))]
-    pub fn into_memory(self) -> M {
-        self.mem
-    }
-
-    #[cfg(test)]
-    pub fn try_write_frame(&self, frame: &[u8]) -> bool {
-        let Some(frame_size) = NonZeroUsize::new(frame.len()) else {
-            return false;
-        };
-        let Ok(mut frame_mut) = self.claim_frame(frame_size) else {
-            return false;
-        };
-        frame_mut.copy_from_slice(frame);
-        frame_mut.finish();
-        true
-    }
-}
-
-/// An exclusively owned, claimed-but-unpublished frame.
-///
-/// [`FrameMut::finish`] commits the frame; it is the only way to make the
-/// payload visible to the receiver. Dropping the frame instead abandons
-/// the claim: the slot stays unfinished and the receiver ignores it,
-/// exactly as if the writer had died there. A writer that abandons a frame
-/// and still performs the operation it described steps outside the usage
-/// contract — records are published before the recorded operation.
-pub struct FrameMut<'a> {
-    state: SharedState,
-    slot_index: usize,
-    descriptor: u64,
-    content: &'a mut [u8],
-}
-
-impl fmt::Debug for FrameMut<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("FrameMut")
-            .field("slot_index", &self.slot_index)
-            .field("len", &self.content.len())
-            .finish_non_exhaustive()
-    }
-}
-
-impl Deref for FrameMut<'_> {
-    type Target = [u8];
-
-    fn deref(&self) -> &Self::Target {
-        self.content
-    }
-}
-
-impl DerefMut for FrameMut<'_> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.content
-    }
-}
-
-impl FrameMut<'_> {
-    /// Commits the frame, making it visible to the receiver.
-    ///
-    /// If the receiver closed the channel and aborted this frame's slot
-    /// first, the swap fails and the frame is silently discarded: the
-    /// record belongs to the close race and is intentionally excluded
-    /// either way.
-    pub fn finish(self) {
-        // Rule 2: `Release` orders every payload write before the
-        // descriptor.
-        let _ = self.state.table()[self.slot_index].compare_exchange(
-            layout::UNFINISHED,
-            self.descriptor,
-            Ordering::Release,
-            Ordering::Relaxed,
-        );
-    }
-}
-
-// --- The reader side: close and iterate -------------------------------------
-//
-// Closing never waits for writers, and no payload byte is read or copied:
-// the reader keeps the mapping alive and hands out borrows of the validated
-// committed spans on demand. Those borrows are sound because of the
-// protocol, not despite it: a committed span is never written again
-// (committing consumes the writer's frame), every borrow covers exactly one
-// validated committed span, and everything a live writer may still touch —
-// counters, slots, its own claimed or abandoned spans — is disjoint from
-// every committed span. This rests on the constructor contract that the
-// region is accessed only through this protocol; a process scribbling
-// outside the protocol is outside the trust model.
-
-/// Shared-memory metadata that could not have been produced by this
-/// protocol. The region was corrupted; its frames are unusable.
-#[derive(thiserror::Error, Clone, Copy, PartialEq, Eq, Debug)]
-pub enum ProtocolError {
-    #[error("corrupt shared-memory frame descriptor at slot {slot_index}")]
-    CorruptDescriptor { slot_index: usize },
-}
-
-/// A reader over the committed frames of a closed channel, serving them
-/// straight out of the mapping, which stays alive inside this value and
-/// is released when the reader drops. It holds no buffer: iteration
-/// re-reads the frozen descriptor table, so closing allocates nothing.
-pub struct ShmReader<M> {
-    /// Owns the region the views point into; dropped with the reader.
-    #[expect(dead_code, reason = "held to keep the region alive")]
-    mem: M,
-    /// The region views, pointing into the owned region's stable target.
-    state: SharedState,
-    /// Length of the frozen prefix of the descriptor table.
-    slot_count: usize,
-    /// Committed frames in that prefix.
-    frames: usize,
-    complete: bool,
-}
-
-// SAFETY: the reader reads only the header atomics, frozen slots, and
-// immutable committed spans; the stored views point into the mapping's
-// stable, independently owned target, not into the reader value itself.
-unsafe impl<M: Send> Send for ShmReader<M> {}
-// SAFETY: see the `Send` impl.
-unsafe impl<M: Sync> Sync for ShmReader<M> {}
-
-impl<M: AsRawSlice> ShmReader<M> {
-    /// Closes the channel over a shared-memory region and returns the
-    /// reader of its committed frames.
-    ///
-    /// Never blocks on writers: writers admitted before the snapshot race
-    /// per slot, and each raced slot independently ends up committed
-    /// (included) or aborted (excluded). Claims after the snapshot land in
-    /// slots this pass never visits until the CLOSED gate — set before
-    /// this returns — stops them. See the protocol docs at the top of this
-    /// module.
-    ///
-    /// # Safety
-    ///
-    /// Same contract as [`ShmWriter::new`]:
-    ///
-    /// - `mem.as_raw_slice()` must return a stable, valid pointer to the
-    ///   whole region for the reader's lifetime.
-    /// - The region must have been zero-initialized when it was created and
-    ///   accessed only through this protocol since.
-    ///
-    /// # Errors
-    ///
-    /// [`ProtocolError`] when the shared-memory metadata could not have
-    /// been produced by a correct writer; the region was corrupted and its
-    /// frames are unusable.
-    ///
-    /// # Panics
-    ///
-    /// Panics when the region is not `u64`-aligned or its size is outside
-    /// the supported range (see [`SharedState::new`]).
-    pub unsafe fn close(mem: M) -> Result<Self, ProtocolError> {
-        // SAFETY: forwarded from this function's contract, which keeps the
-        // region valid for the reader's lifetime — and so for every use of
-        // the views, which are stored in and dropped with the reader.
-        let state = unsafe { SharedState::new(mem.as_raw_slice()) };
-        let slot_count;
-        let mut frames = 0;
-        let complete;
-        {
-            // The close boundary (rule 1): claims at or before this
-            // snapshot are inside it, later ones land in slots this pass
-            // never visits. The count is clamped to the table capacity, so
-            // a counter inflated by failed claims (or by a foreign
-            // scribble) degrades to a full-table sweep, not an error.
-            let claims = state.header().claims.load(Ordering::Relaxed);
-            slot_count =
-                usize::try_from(claims & !CLOSED).unwrap_or(usize::MAX).min(state.table().len());
-            // The same load carries the completeness verdict: a gate set
-            // before this boundary is a failed claim's loss report — or an
-            // earlier close, and a re-close cannot vouch for records
-            // refused since then (rule 1).
-            complete = claims & CLOSED == 0;
-
-            // Gate further claims, so stragglers stop claiming (and
-            // materializing pages) forever. Cheap: the creator pre-faulted
-            // this page where first touches are expensive. Claims racing
-            // between the snapshot and this gate are dropped soundly (see
-            // the module docs above).
-            state.header().claims.fetch_or(CLOSED, Ordering::Relaxed);
-
-            // Freeze pass: drive every admitted slot to a terminal state
-            // and validate the committed descriptors. After this loop the
-            // snapshot's slice of the descriptor table can no longer change
-            // — late writers lose their commit race against `ABORTED` — so
-            // iteration re-reads the table instead of snapshotting it:
-            // nothing is copied or allocated.
-            for slot_index in 0..slot_count {
-                // Rule 3: `Acquire` on failure makes a committed payload
-                // visible.
-                let Err(bits) = state.table()[slot_index].compare_exchange(
-                    layout::UNFINISHED,
-                    layout::ABORTED,
-                    Ordering::AcqRel,
-                    Ordering::Acquire,
-                ) else {
-                    // The receiver won the race: the unfinished slot is now
-                    // aborted and stays ignored.
-                    continue;
-                };
-                if bits == layout::ABORTED {
-                    // Aborted by an earlier close over the same region;
-                    // still ignored.
-                    continue;
-                }
-                // Any other terminal value must be a committed descriptor
-                // with a valid span; a foreign scribble fails the decode.
-                if layout::decode(state.len, bits).is_none() {
-                    return Err(ProtocolError::CorruptDescriptor { slot_index });
-                }
-                frames += 1;
-            }
-        }
-
-        Ok(Self { mem, state, slot_count, frames, complete })
-    }
-
-    /// Iterates over the committed frames in claim order.
-    pub fn iter(&self) -> Iter<'_> {
-        self.into_iter()
-    }
-
-    /// Whether every record a writer published made it in.
-    ///
-    /// False when a claim failed before the channel closed — the region
-    /// was out of space, or a frame exceeded the frame limit: its record
-    /// was lost, and the frames under-report what writers went on to do.
-    /// Consumers that need completeness must reject them.
-    #[must_use]
-    pub const fn is_complete(&self) -> bool {
-        self.complete
-    }
-}
-
-impl<M> fmt::Debug for ShmReader<M> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("ShmReader")
-            .field("frames", &self.frames)
-            .field("complete", &self.complete)
-            .finish_non_exhaustive()
-    }
-}
-
-/// Iterator over a [`ShmReader`]'s committed frames, in claim order.
-pub struct Iter<'a> {
-    /// Base address of the mapping the descriptors' spans point into.
-    base: *const u8,
-    /// The not-yet-visited part of the table's frozen prefix.
-    table: &'a [AtomicU64],
-    /// Mapping length, which decoding validates descriptors against.
-    mapping_len: usize,
-    /// Committed frames not yet yielded.
-    remaining: usize,
-}
-
-impl<'a> Iterator for Iter<'a> {
-    type Item = &'a [u8];
-
-    fn next(&mut self) -> Option<Self::Item> {
-        while let Some((slot, rest)) = self.table.split_first() {
-            self.table = rest;
-            // The slot is terminal (`close` froze it), so this plain load
-            // reads the same value the freeze pass saw; the transfer that
-            // carried the reader to this thread carried the freeze pass's
-            // `Acquire` payload visibility with it (rule 3).
-            let bits = slot.load(Ordering::Relaxed);
-            // `None` is an aborted slot: nothing was published. Corrupt
-            // values cannot appear — `close` already failed the channel on
-            // them — so every decoded span is one `close` validated.
-            let Some(span) = layout::decode(self.mapping_len, bits) else {
-                continue;
-            };
-            self.remaining -= 1;
-            // SAFETY: `close` validated the span against the mapping's
-            // layout, and a committed span is immutable for the mapping's
-            // lifetime (see the section comment above); the reader
-            // borrowed for `'a` keeps the mapping alive and mapped.
-            return Some(unsafe { slice::from_raw_parts(self.base.add(span.offset), span.len) });
-        }
-        None
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (self.remaining, Some(self.remaining))
-    }
-}
-
-impl<'a, M: AsRawSlice> IntoIterator for &'a ShmReader<M> {
-    type IntoIter = Iter<'a>;
-    type Item = &'a [u8];
-
-    fn into_iter(self) -> Iter<'a> {
-        Iter {
-            base: self.state.base(),
-            table: &self.state.table()[..self.slot_count],
-            mapping_len: self.state.len,
-            remaining: self.frames,
-        }
-    }
-}
-
 /// Materializes the page backing the protocol header without changing
 /// protocol state, so that neither a writer's first claim nor
 /// [`ShmReader::close`]'s snapshot pays for the backing file's first
@@ -720,7 +134,7 @@ impl<'a, M: AsRawSlice> IntoIterator for &'a ShmReader<M> {
 #[cfg(target_os = "linux")]
 pub unsafe fn pre_fault(mem: &impl AsRawSlice) {
     // SAFETY: forwarded from this function's contract.
-    let state = unsafe { SharedState::new(mem.as_raw_slice()) };
+    let mapped = unsafe { MappedLayout::new(mem.as_raw_slice()) };
     // A compare-exchange of zero with zero on the claim counter: on an
     // untouched region it performs a real write — allocating the first
     // block of a sparse backing file — without changing protocol state. If
@@ -728,7 +142,7 @@ pub unsafe fn pre_fault(mem: &impl AsRawSlice) {
     // exchange changes nothing. (An `or` of zero would not do: the
     // compiler may lower it to a plain load, which materializes only a
     // hole page without allocating the block.)
-    let _ = state.header().claims.compare_exchange(0, 0, Ordering::Relaxed, Ordering::Relaxed);
+    let _ = mapped.header().claims.compare_exchange(0, 0, Ordering::Relaxed, Ordering::Relaxed);
 }
 
 #[cfg(test)]
